requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
pbr==6.0.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
Tempita==0.5.2
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
